source/requirements.txt
inflection
email-validator
fastapi
fastapi-pagination
httpx
mimesis
psycopg2-binary
pydantic
pytest
python-dotenv
sqlalchemy
sqlmodel
requests
uvicorn